worked furher on bulk export
There's a rate limit in the Discord API but i didn't get when it occurs... tried several things but nothing really worked. Now when the rate limit becomes effective, i just wait three times as long and then try again...
also used auto format to let it look a bit better and added feedback when the export is ready.