Strings Update, improved Support for copying to Clipboard
Added Support for Copying to Clipboard in Bulk Modify.
Changed Behavior of manually Copying to Clipboard to also add the Location to done even though the user might not have pasted it to Discord.
Adapted Strings to better reflect if a Command will be copied to Clipboard or sent via WebHook.
Unified style to write Poi, WebHook and Pogo Tools whic was mixed before.