Fixed selections showing up multiple times
the portalSelected Hook is called two or three times when selecting a new portal. this caused the radio buttons to get added two or more times. this is now fixed by changing the hook from portalSelected to portalDetailsUpdated and removing the hook as long as the buttons wait to get added.